Using update from temp table

+++ crates/db_schema/replaceable_schema/utils.sql
@@ -154,7 +154,7 @@ $a$;
 -- Edit community aggregates to include voters as active users
 CREATE OR REPLACE FUNCTION r.community_aggregates_activity (i text)
     RETURNS TABLE (
-        count_ integer,
+        count_ bigint,
         community_id_ integer)
     LANGUAGE plpgsql
     AS $$
